requirements.txt
absl-py==1.2.0
addict==2.4.0
albumentations==1.2.1
antlr4-python3-runtime==4.9.3
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
black==21.4b2
bleach==5.0.1
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.0.4
cityscapesScripts==2.2.0
click==8.1.3
cloudpickle==2.2.0
codecov==2.1.12
colorama==0.4.4
coloredlogs==15.0.1
commonmark==0.9.1
conda==22.9.0
conda-content-trust==0+unknown
conda-package-handling==1.8.1
coverage==6.4.4
cryptography==36.0.0
cycler==0.11.0
Cython==0.29.28
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
detectron2==0.6+cu111
docker-pycreds==0.4.0
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
flake8==5.0.4
flatbuffers==22.9.24
fonttools==4.37.1
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.49.1
h5py==3.7.0
humanfriendly==10.0
hydra-core==1.2.0
idna==3.3
imagecodecs==2021.8.26
imageio==2.21.3
importlib-metadata==4.12.0
iniconfig==1.1.1
interrogate==1.5.0
iopath==0.1.9
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jdet==0.2.0.0
jedi==0.18.1
Jinja2==3.1.2
jittor==1.3.5.16
joblib==1.2.0
json5==0.9.10
jsonschema==4.15.0
jupyter==1.0.0
jupyter_client==7.3.5
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.15.1
jupyterlab-widgets==3.0.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lab==7.1
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mindspore-gpu==1.8.1
mistune==2.0.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcls==0.23.2
mmcv-full==1.4.2
mmdet==2.25.2
mmsegmentation==0.27.0
model-index==0.1.11
mypy-extensions==0.4.3
nbclassic==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.6
nibabel==4.0.2
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.21.5
oauthlib==3.2.1
omegaconf==2.2.3
opencv-contrib-python==4.6.0.66
opencv-python==3.4.13.47
opencv-python-headless==4.6.0.66
openmim==0.3.1
openpsg==0.5.0
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
panopticapi==0.1
parso==0.8.3
pathspec==0.10.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==22.2.2
pluggy==1.0.0
portalocker==2.5.1
prettytable==3.4.1
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.5
pycodestyle==2.9.1
pycosat==0.6.3
pycparser==2.21
pydot==1.4.2
pyflakes==2.5.0
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
qudida==0.0.4
regex==2022.9.13
requests==2.27.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
ruamel-yaml-conda==0.15.100
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
Send2Trash==1.8.0
sentry-sdk==1.9.10
setproctitle==1.3.2
setuptools==61.2.0
Shapely==1.8.4
shortuuid==1.0.9
SimpleITK==2.2.0
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.0
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
terminado==0.15.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.4.12
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado==6.2
tqdm==4.63.0
traitlets==5.3.0
txt2tags==3.7
typing==3.7.4.3
typing_extensions==4.3.0
urllib3==1.26.12
wandb==0.13.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==4.0.3
wrapt==1.14.1
xdoctest==1.0.2
xmltodict==0.13.0
yacs==0.1.8
yapf==0.32.0
zipp==3.8.1
